+++ tests/lib/Botkit.test.js
