+++ tests/lib/Slackbot_worker.test.js
